requirements.txt
Flask==2.1.2
requests==2.28.1
python-dotenv==0.21.0
overpy==0.6
psycopg2-binary==2.9.3
numpy==1.23.5